requirements.txt
prefect>=3.1,<4
pandas==2.2.*
numpy==1.26.*

SQLAlchemy==2.0.*
psycopg2-binary==2.9.*

pyarrow==16.*

gcsfs==2024.6.*
google-cloud-storage==2.16.*
google-cloud-secret-manager==2.20.*
google-auth==2.30.*

python-dotenv==1.0.*
tenacity==8.3.*

GitPython